+++ app/src/main/java/com/example/vcubmonitor/json/VolleyResultCallBack.kt
@@ -1,0 +1,8 @@
+package com.example.vcubmonitor.json
+
+import com.android.volley.VolleyError
+
+interface VolleyResultCallBack {
+    fun onVolleyResultListener(response: Any?)
+    fun onVolleyErrorListener(error :Any?)
+}
